+++ internal/evaluator/tests/evaluator_test.go
@@ -1,6 +1,7 @@
-package evaluator
+package tests
 
 import (
+	"interpreter/internal/evaluator"
 	"strings"
 	"testing"
 
@@ -46,7 +47,7 @@ func testEval(t *testing.T, source string) object.Object {
 	// 1. Arrange
 	l := lexer.NewLexer(strings.NewReader(source))
 	p := parser.NewParser(l)
-	v := NewASTVisitor()
+	v := evaluator.NewASTVisitor()
 
 	// 2. Act
 	program := p.Parse()
@@ -55,7 +56,7 @@ func testEval(t *testing.T, source string) object.Object {
 	require.Len(t, p.Errors(), 0)
 	require.Greater(t, len(program.Statements), 0)
 
-	return EvaluateStatement(program, v)
+	return evaluator.EvaluateStatement(program, v)
 }
 
 func testIntegerObject(t *testing.T, obj object.Object, expected int64) {
@@ -253,59 +254,6 @@ func TestLetStatement(t *testing.T) {
 			testIntegerObject(t, got, tt.expected)
 		})
 	}
-}
-
-func TestFunctionObject(t *testing.T) {
-	// 1. Arrange
-	source := "fn(x) { x + 2; };"
-
-	// 2. Act
-	got := testEval(t, source)
-	fn, ok := got.(*object.Function)
-	require.True(t, ok, "expected function, but got=%T", got)
-	require.Len(t, fn.Args, 1, "expected one parameter in function")
-	require.Equal(t, "x", fn.Args[0].String(), "parameter is not x")
-	require.Equal(t, "(x + 2)", fn.Body.String())
-}
-
-func TestFunctionApplication(t *testing.T) {
-	for _, tt := range []struct {
-		source   string
-		expected int64
-	}{
-		{"let identity = fn(x) { x; }; identity(5);", 5},
-		{"let identity = fn(x) { return x; }; identity(5);", 5},
-		{"let double = fn(x) { x * 2; }; double(5);", 10},
-		{"let add = fn(x, y) { x + y; }; add(5, 5);", 10},
-		{"let add = fn(x, y) { x + y; }; add(5 + 5, add(5, 5));", 20},
-		{"fn(x) { x; }(5)", 5},
-	} {
-		t.Run(tt.source, func(t *testing.T) {
-			// 1. Arrange
-			got := testEval(t, tt.source)
-
-			// 2. Act
-			testIntegerObject(t, got, tt.expected)
-		})
-	}
-}
-
-func TestClosures(t *testing.T) {
-	// 1. Arrange
-	source := `
-let newAdder = fn(x) {
-	fn(y) { x + y };
-};
-
-let addTwo = newAdder(2);
-addTwo(2);
-`
-
-	// 2. Act
-	got := testEval(t, source)
-
-	// 3. Assert
-	testIntegerObject(t, got, 4)
 }
 
 func TestEvaluateStringLiterals(t *testing.T) {
